Fixed Task and Supplier Pages, both now work as intended

+++ SoftStocks/SoftStocksGUI/Widgets/frmSupplierEntry.cs
@@ -19,7 +19,7 @@ namespace SoftStocksGUI.Widgets
 
 		private int entryId;
 
-		public frmSupplierEntry(string name = "", string contact = "", string purchases = "", string address = "", int id = 0)
+		public frmSupplierEntry(int id = -1, string name = "Example Name", string contact = "Example Number", string purchases = "Example Purchases", string address = "Example Address")
 		{
 			InitializeComponent();
 
@@ -29,7 +29,40 @@ namespace SoftStocksGUI.Widgets
 			this.lblSupplierContactNumberEntry.Text = contact;
 			this.lblSupplierContactNameEntry.Text = purchases;
 			this.lblSupplierAddressEntry.Text = address;
-			this.lblSupplierId.Text = $"{id}";
+			this.lblSupplierId.Text = $"{entryId}";
+
+			if (entryId == -1)
+			{
+				using (SoftStocksDBContext db = new SoftStocksDBContext())
+				{
+					var namesQuery = db.Suppliers.SingleOrDefault(s => s.Name == this.lblSupplierNameEntry.Text);
+					if (namesQuery == null)
+					{
+						Supplier newSupplier = new Supplier
+						{
+							Name = this.lblSupplierNameEntry.Text,
+							ContactNumber = this.lblSupplierContactNumberEntry.Text,
+							PrimaryContact = this.lblSupplierContactNameEntry.Text,
+							BusinessAddress = this.lblSupplierAddressEntry.Text
+						};
+
+						db.Suppliers.Add(newSupplier);
+						db.SaveChanges();
+
+						var row = db.Suppliers.Single(s => s.Name == name);
+						if (row != null)
+						{
+							entryId = row.Id;
+							this.lblSupplierId.Text = $"{entryId}";
+						}
+					}
+					else
+					{
+						MessageBox.Show("Must First Change Example Entry!");
+						this.Close();
+					}
+				}
+			}
 
 		}
 
@@ -43,54 +76,43 @@ namespace SoftStocksGUI.Widgets
 
 			using (SoftStocksDBContext db = new SoftStocksDBContext())
 			{
-				var row = db.Suppliers.FirstOrDefault(s => s.Id == entryId);
-				if (row != null)
+
+				//Check For Same Names
+				var namesQuery = db.Suppliers.SingleOrDefault(s => s.Name == this.lblSupplierNameEntry.Text);
+				if (namesQuery == null)
 				{
-					row.Name = this.lblSupplierNameEntry.Text;
-					row.ContactNumber = this.lblSupplierContactNumberEntry.Text;
-					row.PrimaryContact = this.lblSupplierContactNameEntry.Text;
-					row.BusinessAddress = this.lblSupplierAddressEntry.Text;
+					var row = db.Suppliers.FirstOrDefault(s => s.Id == entryId);
+					if (row != null)
+					{
+						row.Name = this.lblSupplierNameEntry.Text;
+						row.ContactNumber = this.lblSupplierContactNumberEntry.Text;
+						row.PrimaryContact = this.lblSupplierContactNameEntry.Text;
+						row.BusinessAddress = this.lblSupplierAddressEntry.Text;
+					}
+					else
+					{
+						Supplier newSupplier = new Supplier
+						{
+							Name = this.lblSupplierNameEntry.Text,
+							ContactNumber = this.lblSupplierContactNumberEntry.Text,
+							PrimaryContact = this.lblSupplierContactNameEntry.Text,
+							BusinessAddress = this.lblSupplierAddressEntry.Text
+						};
+						db.Suppliers.Add(newSupplier);
+					}
+					db.SaveChanges();
 				}
 				else
 				{
-					Supplier newSupplier = new Supplier
-					{
-						Name = this.lblSupplierNameEntry.Text,
-						ContactNumber = this.lblSupplierContactNumberEntry.Text,
-						PrimaryContact = this.lblSupplierContactNameEntry.Text,
-						BusinessAddress = this.lblSupplierAddressEntry.Text
-					};
-					db.Suppliers.Add(newSupplier);
+					MessageBox.Show("Name Already In Use! Change Name And Try Again.");
 				}
 
-				db.SaveChanges();
+
 			}
 
 		}
 
-
-		private void lblSupplierNameEntry_TextChanged(object sender, EventArgs ge)
-		{
-			removeRowifBlank();
-		}
-
-		private void lblSupplierContactNumberEntry_TextChanged(object sender, EventArgs e)
-		{
-			removeRowifBlank();
-		}
-
-
-		private void lblSupplierContactNameEntry_TextChanged(object sender, EventArgs e)
-		{
-			removeRowifBlank();
-		}
-
-		private void lblSupplierAddressEntry_TextChanged(object sender, EventArgs e)
-		{
-			removeRowifBlank();
-		}
-
-		private void removeRowifBlank()
+		private void removeRowifBlank(object sender, EventArgs e)
 		{
 			if (lblSupplierNameEntry.Text == String.Empty && lblSupplierContactNumberEntry.Text == String.Empty && lblSupplierContactNameEntry.Text == String.Empty && lblSupplierAddressEntry.Text == String.Empty)
 			{
@@ -102,12 +124,20 @@ namespace SoftStocksGUI.Widgets
 		{
 			using (SoftStocksDBContext db = new SoftStocksDBContext())
 			{
-				var supplierRow = new Supplier { Id = entryId };
-				db.Suppliers.Attach(supplierRow);
-				db.Suppliers.Remove(supplierRow);
-				db.SaveChanges();
-			}
 
+				//db.Suppliers.RemoveAll(s => s.Id == entryId);
+				//var supplierRow = new Supplier { Id = entryId };
+				Supplier supplierRow = db.Suppliers.Find(entryId);
+				if (supplierRow != null)
+				{
+					db.Suppliers.Remove(supplierRow);
+					db.SaveChanges();
+				}
+				else
+				{
+					MessageBox.Show($"Could not delete Supplier, could not find id: {entryId}");
+				}
+			}
 			this.Close();
 		}
 	}
